Only highlight type signatures split across lines
This will properly highlight type signatures that are split across
lines. The function name will not be highlighted, however, if the
double-colon (`::`) is not on the same line. There doesn't seem
to be a way to fix this that I can think of. Foreign imports can
still, however, be split over multiple lines.